Completed laboratory work No. 10:
Task 1.
Write a method that reads a text file and returns it as a list of strings.

Task 2.
Write a method that writes the string passed as a parameter to a file.

Task 3.
Using solution 1 and 2, write a method that merges two text files into one.

Task 4.
Write a method that replaces everything in a file except letters and numbers with the ‘$’ sign